+++ Source/WinCopies.Collections.Shared/Generic/CountableEnumerableArray.cs
@@ -16,6 +16,7 @@
  * along with the WinCopies Framework.  If not, see <https://www.gnu.org/licenses/>. */
 
 using System.Collections;
+using System.Collections.Generic;
 
 #if WinCopies2
 using WinCopies.Collections.DotNetFix;
@@ -24,31 +25,75 @@ using WinCopies.Collections.DotNetFix.Generic;
 
 namespace WinCopies.Collections.Generic
 {
-    public class CountableEnumerableArray<T> : ICountableEnumerable<T>
+    public class UIntCountableEnumerator<T> : Enumerator<T, ICountableEnumeratorInfo<T>, T, IUIntCountableEnumeratorInfo<T>>, IUIntCountableEnumeratorInfo<T>
     {
-        private T[] _array;
+        public uint Count => (uint)InnerEnumerator.Count;
 
-        public CountableEnumerableArray(T[] array) => _array = array;
+        protected override T CurrentOverride => InnerEnumerator.Current;
 
-        public int Count => _array.Length;
+        public override bool? IsResetSupported => InnerEnumerator.IsResetSupported;
 
-        public System.Collections.Generic.IEnumerator<T> GetEnumerator() => new ArrayEnumerator<T>(_array);
+        public UIntCountableEnumerator(ICountableEnumeratorInfo<T> enumerator) : base(enumerator) { /* Left empty. */ }
+
+        protected override bool MoveNextOverride() => InnerEnumerator.MoveNext();
+    }
+
+    public interface IReadOnlyList<out T> : System.Collections.Generic.IReadOnlyList<T>, ICountableEnumerable<T>
+    {
+        // Left empty.
+    }
+
+    public class CountableEnumerableArray<T> : IReadOnlyList<T>
+    {
+        protected T[] Array { get; }
+
+        public int Count => Array.Length;
+
+        public T this[int index] => Array[index];
+
+        public CountableEnumerableArray(in T[] array) => Array = array;
+
+        public
+#if WinCopies3
+            ICountableEnumeratorInfo<T>
+#else
+         System.Collections.Generic.IEnumerator<T> 
+#endif
+            GetEnumerator() => new ArrayEnumerator<T>(Array);
 
         System.Collections.IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+#if WinCopies3
+        ICountableEnumerator<T> ICountableEnumerable<T>.GetEnumerator() => GetEnumerator();
+
+        System.Collections.Generic.IEnumerator<T> System.Collections.Generic.IEnumerable<T>.GetEnumerator() => GetEnumerator();
+#endif
     }
 
     public class UIntCountableEnumerableArray<T> : IUIntCountableEnumerable<T>
     {
-        private CountableEnumerableArray<T> _array;
+        private readonly CountableEnumerableArray<T> _array;
 
-        public UIntCountableEnumerableArray(T[] array) : this(new CountableEnumerableArray<T>(array)) { /* Left empty. */ }
+        public UIntCountableEnumerableArray(in T[] array) : this(new CountableEnumerableArray<T>(array)) { /* Left empty. */ }
 
-        public UIntCountableEnumerableArray(CountableEnumerableArray<T> array) => _array = array;
+        public UIntCountableEnumerableArray(in CountableEnumerableArray<T> array) => _array = array;
 
         public uint Count => (uint)_array.Count;
 
-        public System.Collections.Generic.IEnumerator<T> GetEnumerator() => _array.GetEnumerator();
+        public
+#if WinCopies3
+            IUIntCountableEnumeratorInfo<T>
+#else
+         System.Collections.Generic.IEnumerator<T> 
+#endif
+            GetEnumerator() => new UIntCountableEnumerator<T>(_array.GetEnumerator());
 
-        System.Collections.IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable)_array).GetEnumerator();
+#if WinCopies3
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+        IUIntCountableEnumerator<T> IUIntCountableEnumerable<T>.GetEnumerator() => GetEnumerator();
+
+        System.Collections.Generic.IEnumerator<T> System.Collections.Generic.IEnumerable<T>.GetEnumerator() => GetEnumerator();
+#endif
     }
 }